[Recharge] Update more amount data type to Long in promo checkout common

+++ features/promo/promo_checkout_common/src/main/java/com/tokopedia/promocheckout/common/domain/model/CheckUmrahPromoCodeData.kt
@@ -27,7 +27,7 @@ data class CheckUmrahPromoCodeData(
         var cashbackWalletAmount : Long = 0L,
         @SerializedName("cashbackAdvocateReferralAmount")
         @Expose
-        var cashbackAdvocateReferralAmount : Int = 0,
+        var cashbackAdvocateReferralAmount : Long = 0L,
         @SerializedName("invoiceDescription")
         @Expose
         var invoiceDescription : String = "",
